Refactor: move prompts into utils
